Add instructions for how to run the pipelines
Other than the instructions, the modified version of the alevin-fry is deleted since that version is not used in our benchmarkings. Instead, we have used the tagged 0.1.0 release with the \texttt{rand} crate additionally pinned at 0.7.3 to enable compilation.
Some issues with the scripts for gathering the data and references are also addressed.